Move field data interfaces for exponential histogram to core.
