+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/exponentialhistogram/fielddata/LeafExponentialHistogramFieldData.java
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-package org.elasticsearch.xpack.exponentialhistogram.fielddata;
+package org.elasticsearch.xpack.core.exponentialhistogram.fielddata;
 
 import org.elasticsearch.index.fielddata.LeafFieldData;
