+++ x-pack/plugin/mapper-exponential-histogram/src/main/java/org/elasticsearch/xpack/exponentialhistogram/fielddata/IndexExponentialHistogramFieldData.java
@@ -9,6 +9,7 @@ package org.elasticsearch.xpack.exponentialhistogram.fielddata;
 
 import org.elasticsearch.index.fielddata.IndexFieldData;
 import org.elasticsearch.search.aggregations.support.ValuesSourceType;
+import org.elasticsearch.xpack.core.exponentialhistogram.fielddata.LeafExponentialHistogramFieldData;
 import org.elasticsearch.xpack.exponentialhistogram.aggregations.support.ExponentialHistogramValuesSourceType;
 
 /**
